+++ ocred/preprocessing.py
@@ -7,6 +7,9 @@ import numpy.typing as npt
 from scipy import ndimage
 from skimage.filters import threshold_local
 
+_dep_warn_inplace = "inplace is deprecated and was removed in v0.3.0; Preprocessor now alters self.img directly"
+_dep_warn_overriden_image = "overriden_image is deprecated and was removed in v0.3.0; Preprocessor now only alters self.img"
+
 
 class Preprocessor:
     """
@@ -17,31 +20,28 @@ class Preprocessor:
             Path of the image or a numpy array.
 
     Examples:
+        >>> import sys
+        >>> sys.displayhook = lambda x: None
         >>> import cv2
         >>> from scipy import ndimage
         >>> from ocred import Preprocessor
         >>> # scan the image and copy the scanned image
         >>> preprocessed = Preprocessor("images/CosmosTwo.jpg")
         >>> # scan the image and copy the scanned image
-        >>> scanned = preprocessed.scan(inplace=True)
-        >>> orig = scanned.copy()
+        >>> preprocessed.scan()
+        >>> orig = preprocessed.img.copy()
         >>> # remove noise
-        ... noise_free = preprocessed.remove_noise(
-        ...     inplace=True, overriden_image=scanned
-        ... )
+        >>> preprocessed.remove_noise()
         >>> # thicken the ink to draw Hough lines better
-        >>> thickened = preprocessed.thicken_font(
-        ...     inplace=True, overriden_image=noise_free
-        ... )
+        >>> preprocessed.thicken_font()
         >>> # calculate the median angle of all the Hough lines
-        >>> _, median_angle = preprocessed.rotate(
-        ...     inplace=True, overriden_image=thickened
-        ... )
+        >>> _, median_angle = preprocessed.rotate()
         >>> # rotate the original scanned image
         >>> rotated = ndimage.rotate(orig, median_angle)
         >>> # remove noise again
-        >>> final_img = preprocessed.remove_noise(inplace=True, overriden_image=rotated)
-        >>> cv2.imwrite("preprocessed.png", final_img)
+        >>> preprocessed = Preprocessor(rotated)
+        >>> preprocessed.remove_noise()
+        >>> cv2.imwrite("preprocessed.png", preprocessed.img)
         True
     """
 
@@ -58,7 +58,7 @@ class Preprocessor:
         self,
         *,
         save: Optional[bool] = False,
-        inplace: Optional[bool] = False,
+        inplace: Optional[Union[bool, None]] = None,
         iterations: Optional[int] = 1,
         overriden_image: Union[
             npt.NDArray[np.int64], npt.NDArray[np.float64], None
@@ -70,39 +70,41 @@ class Preprocessor:
         Args:
             save:
                 Saves the resultant image.
-            inplace:
-                Edits the image inplace.
             iterations:
                 Number of times the image is processed.
+            inplace:
+                DANGER: Deprecated since version v0.3.0.
+                Was intended to edit the image inplace, but never actually worked.
             overriden_image:
-                Overrides the image passes through the constructor.
+                DANGER: Deprecated since version v0.3.0.
+                Was used to pass a new image to the method but was redundant and buggy.
 
         Returns:
             noise_free_image:
                 The noise free image.
         """
-        if not inplace:
-            img = self.img.copy() if overriden_image is None else overriden_image.copy()
-        else:
-            img = self.img if overriden_image is None else overriden_image
+        if inplace is not None:
+            raise DeprecationWarning(_dep_warn_inplace)
+        if overriden_image is not None:
+            raise DeprecationWarning(_dep_warn_overriden_image)
 
         kernel: Union[npt.NDArray[np.int64]] = np.ones((1, 1), np.uint8)
-        img = cv2.dilate(img, kernel, iterations=iterations)
+        self.img = cv2.dilate(self.img, kernel, iterations=iterations)
         kernel = np.ones((1, 1), np.uint8)
-        img = cv2.erode(img, kernel, iterations=iterations)
-        img = cv2.morphologyEx(img, cv2.MORPH_CLOSE, kernel)
-        img = cv2.medianBlur(img, 3)
+        self.img = cv2.erode(self.img, kernel, iterations=iterations)
+        self.img = cv2.morphologyEx(self.img, cv2.MORPH_CLOSE, kernel)
+        self.img = cv2.medianBlur(self.img, 3)
 
         if save:
-            cv2.imwrite("noise_free.png", img)
+            cv2.imwrite("noise_free.png", self.img)
 
-        return img
+        return self.img
 
     def thicken_font(
         self,
         *,
         save: Optional[bool] = False,
-        inplace: Optional[bool] = False,
+        inplace: Optional[Union[bool, None]] = None,
         iterations: Optional[int] = 2,
         overriden_image: Union[
             npt.NDArray[np.int64], npt.NDArray[np.float64], None
@@ -114,37 +116,39 @@ class Preprocessor:
         Args:
             save:
                 Saves the resultant image.
-            inplace:
-                Edits the image inplace.
             iterations:
                 Number of times the image is processed.
+            inplace:
+                DANGER: Deprecated since version v0.3.0.
+                Was intended to edit the image inplace, but never actually worked.
             overriden_image:
-                Overrides the image passes through the constructor.
+                DANGER: Deprecated since version v0.3.0.
+                Was used to pass a new image to the method but was redundant and buggy.
 
         Returns:
             thickened_image:
                 The thickened image.
         """
-        if not inplace:
-            img = self.img.copy() if overriden_image is None else overriden_image.copy()
-        else:
-            img = self.img if overriden_image is None else overriden_image
+        if inplace is not None:
+            raise DeprecationWarning(_dep_warn_inplace)
+        if overriden_image is not None:
+            raise DeprecationWarning(_dep_warn_overriden_image)
 
-        img = cv2.bitwise_not(img)
+        self.img = cv2.bitwise_not(self.img)
         kernel: Union[npt.NDArray[np.int64]] = np.ones((2, 2), np.uint8)
-        img = cv2.dilate(img, kernel, iterations=iterations)
-        img = cv2.bitwise_not(img)
+        self.img = cv2.dilate(self.img, kernel, iterations=iterations)
+        self.img = cv2.bitwise_not(self.img)
 
         if save:
-            cv2.imwrite("thick_font.png", img)
+            cv2.imwrite("thick_font.png", self.img)
 
-        return img
+        return self.img
 
     def scan(
         self,
         *,
         save: Optional[bool] = False,
-        inplace: Optional[bool] = False,
+        inplace: Optional[Union[bool, None]] = None,
         overriden_image: Union[
             npt.NDArray[np.int64], npt.NDArray[np.float64], None
         ] = None
@@ -156,33 +160,35 @@ class Preprocessor:
             save:
                 Saves the resultant image.
             inplace:
-                Edits the image inplace.
+                DANGER: Deprecated since version v0.3.0.
+                Was intended to edit the image inplace, but never actually worked.
             overriden_image:
-                Overrides the image passes through the constructor.
+                DANGER: Deprecated since version v0.3.0.
+                Was used to pass a new image to the method but was redundant and buggy.
 
         Returns:
             scanned_image:
                 The scanned image.
         """
-        if not inplace:
-            img = self.img.copy() if overriden_image is None else overriden_image.copy()
-        else:
-            img = self.img if overriden_image is None else overriden_image
+        if inplace is not None:
+            raise DeprecationWarning(_dep_warn_inplace)
+        if overriden_image is not None:
+            raise DeprecationWarning(_dep_warn_overriden_image)
 
-        img = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
-        thr = threshold_local(img, 11, offset=10, method="gaussian")
-        img = (img > thr).astype("uint8") * 255
+        self.img = cv2.cvtColor(self.img, cv2.COLOR_BGR2GRAY)
+        thr = threshold_local(self.img, 11, offset=10, method="gaussian")
+        self.img = (self.img > thr).astype("uint8") * 255
 
         if save:
-            cv2.imwrite("scanned.png", img)
+            cv2.imwrite("scanned.png", self.img)
 
-        return img
+        return self.img
 
     def rotate(
         self,
         *,
         save: Optional[bool] = False,
-        inplace: Optional[bool] = False,
+        inplace: Optional[Union[bool, None]] = None,
         overriden_image: Union[
             npt.NDArray[np.int64], npt.NDArray[np.float64], None
         ] = None
@@ -194,9 +200,11 @@ class Preprocessor:
             save:
                 Saves the resultant image.
             inplace:
-                Edits the image inplace.
+                DANGER: Deprecated since version v0.3.0.
+                Was intended to edit the image inplace, but never actually worked.
             overriden_image:
-                Overrides the image passes through the constructor.
+                DANGER: Deprecated since version v0.3.0.
+                Was used to pass a new image to the method but was redundant and buggy.
 
         Returns:
             rotated_image:
@@ -204,12 +212,12 @@ class Preprocessor:
             median_angle:
                 The angly by which it is rotated.
         """
-        if not inplace:
-            img = self.img.copy() if overriden_image is None else overriden_image.copy()
-        else:
-            img = self.img if overriden_image is None else overriden_image
+        if inplace is not None:
+            raise DeprecationWarning(_dep_warn_inplace)
+        if overriden_image is not None:
+            raise DeprecationWarning(_dep_warn_overriden_image)
 
-        img_edges = cv2.Canny(img, 100, 100, apertureSize=3)
+        img_edges = cv2.Canny(self.img, 100, 100, apertureSize=3)
         lines = cv2.HoughLinesP(
             img_edges,
             rho=1,
@@ -225,9 +233,9 @@ class Preprocessor:
             angles.append(angle)
 
         median_angle = float(np.median(angles))
-        img = ndimage.rotate(img, median_angle)
+        self.img = ndimage.rotate(self.img, median_angle)
 
         if save:
-            cv2.imwrite("rotated.png", img)
+            cv2.imwrite("rotated.png", self.img)
 
-        return img, median_angle
+        return self.img, median_angle